Move to TabExecutor standard
- move file from commands/position/PositionCommandExecutor to commands/PositionCommand
- rename file from WarpCommandExecutor to WarpVersionCommand etc.
