+++ src/main/kotlin/me/marylieh/simplewarp/SimpleWarp.kt
@@ -1,7 +1,6 @@
 package me.marylieh.simplewarp
 
 import me.marylieh.simplewarp.commands.*
-import me.marylieh.simplewarp.commands.position.PositionCommandExecutor
 import me.marylieh.simplewarp.utils.Config
 import me.marylieh.simplewarp.utils.Updater
 import org.bukkit.plugin.java.JavaPlugin
@@ -39,14 +38,14 @@ class SimpleWarp : JavaPlugin() {
         val warpsCommand = getCommand("warps") ?: error("Couldn't get warps command! This should not happen!")
         val warpVersionCommand = getCommand("warpversion") ?: error("Couldn't get warpversion command! This should not happen!")
         val positionCommand = getCommand("position") ?: error("Couldn't get position command! This should not happen!")
-        setWarpCommand.setExecutor(SetWarpCommandExecutor())
-        delWarpCommand.setExecutor(DelWarpCommandExecutor())
-        warpCommand.setExecutor(WarpCommandExecutor())
-        warpsCommand.setExecutor(WarpsCommandExecutor())
-        warpVersionCommand.setExecutor(WarpVersionCommandExecutor())
-        positionCommand.setExecutor(PositionCommandExecutor())
-        warpCommand.setTabCompleter(WarpTabCompleter())
-        delWarpCommand.setTabCompleter(WarpTabCompleter())
+        setWarpCommand.setExecutor(SetWarpCommand())
+        warpsCommand.setExecutor(WarpsCommand())
+        delWarpCommand.setExecutor(DelWarpCommand())
+        warpCommand.setExecutor(WarpCommand())
+        warpCommand.tabCompleter = WarpCommand()
+        warpVersionCommand.setExecutor(WarpVersionCommand())
+        positionCommand.setExecutor(PositionCommand())
+        positionCommand.tabCompleter = PositionCommand()
     }
 
     private fun initConfig() {
